+++ ts-test-w2m/src/app/shared/components/index.ts
@@ -1,0 +1,3 @@
+import { BaseDialogComponent } from './base-dialog/base-dialog.component';
+
+export const SHARED_COMPONENTS = [BaseDialogComponent];