+++ ts-test-w2m/src/app/shared/directives/index.ts
@@ -1,0 +1,3 @@
+import { UppercaseDirective } from './uppercase.directive';
+
+export const DIRECTIVES = [UppercaseDirective];