+++ ts-test-w2m/src/app/store/heroes.actions.ts
@@ -1,6 +1,6 @@
 import { createAction, props } from '@ngrx/store';
 import { HeroesData } from './heroes.model';
-import { Heroe } from '../shared/interfaces/heroes';
+import { Heroe } from '../shared/models/heroes';
 
 export const init = createAction('[Heroes] -  INIT', props<{ data: any }>());
 export const heroesList = createAction('[Heroes] -  LOAD LIST');
@@ -15,7 +15,29 @@ export const addHeroe = createAction(
 
 export const addHeroeSuccess = createAction(
   '[Heroe] -  ADD HEROE SUCCESS',
-  props<{ data: HeroesData }>()
+  props<{
+    data: {
+      heroes: HeroesData;
+      message: string;
+    };
+  }>()
+);
+
+export const updateHeroe = createAction(
+  '[Heroes] -  UPDATE HEROE',
+  props<{
+    data: Heroe;
+  }>()
+);
+
+export const updateHeroeSuccess = createAction(
+  '[Heroe] -  UPDATE HEROE SUCCESS',
+  props<{
+    data: {
+      heroes: HeroesData;
+      message: string;
+    };
+  }>()
 );
 
 export const removeHeroe = createAction(
@@ -27,7 +49,12 @@ export const removeHeroe = createAction(
 
 export const removeHeroeSuccess = createAction(
   '[Heroe] -  REMOVE HEROE SUCCESS',
-  props<{ data: HeroesData }>()
+  props<{
+    data: {
+      heroes: HeroesData;
+      message: string;
+    };
+  }>()
 );
 
 export const findHeroe = createAction(
@@ -41,5 +68,5 @@ export const findHeroe = createAction(
 
 export const findHeroeSuccess = createAction(
   '[Heroe] -  FIND HEROE SUCCESS',
-  props<{ data: HeroesData }>()
+  props<{ data: Heroe }>()
 );