+++ ts-test-w2m/src/app/store/heroes.reducer.ts
@@ -1,9 +1,14 @@
 import { createReducer, on } from '@ngrx/store';
 import { CallState, HeroesData, LoadingState } from './heroes.model';
 import * as Actions from './heroes.actions';
+import { Heroe } from '../shared/models/heroes';
 
 export interface HeroesState {
   data: HeroesData;
+  action: {
+    message: string;
+  };
+  heroeSelected: Heroe;
   callState: CallState;
 }
 
@@ -12,6 +17,14 @@ export const HEROES_INITIAL_STATE: HeroesState = {
     heroes: [],
     pageSize: 1,
     current_page: 1,
+  },
+  heroeSelected: {
+    id: '',
+    firstName: '',
+    lastName: '',
+  },
+  action: {
+    message: '',
   },
   callState: LoadingState.INIT,
 };
@@ -30,14 +43,29 @@ export const heroesReducer = createReducer(
     callState: LoadingState.LOADING,
   })),
 
+  on(Actions.findHeroeSuccess, (state, action) => ({
+    ...state,
+    callState: LoadingState.LOADING,
+    heroeSelected: action.data,
+  })),
+
+  on(Actions.heroesListSuccess, (state, action) => ({
+    ...state,
+    callState: LoadingState.LOADED,
+    data: action.data,
+  })),
+
   on(
-    Actions.heroesListSuccess,
     Actions.addHeroeSuccess,
     Actions.removeHeroeSuccess,
+    Actions.updateHeroeSuccess,
     (state, action) => ({
       ...state,
       callState: LoadingState.LOADED,
-      data: action.data,
+      data: action.data.heroes,
+      action: {
+        message: action.data.message,
+      },
     })
   )
 );