+++ ts-test-w2m/src/app/views/heroes/components/index.ts
@@ -1,0 +1,9 @@
+import { HeroesDetailComponent } from './heroes-detail/heroes-detail.component';
+import { HeroesListComponent } from './heroes-list/heroes-list.component';
+import { HeroesNewComponent } from './heroes-new/heroes-new.component';
+
+export const COMPONENTS = [
+  HeroesListComponent,
+  HeroesDetailComponent,
+  HeroesNewComponent,
+];