+++ ts-test-w2m/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Glossary](#glossary)
 
+## [0.0.2] - 2023-11-14
+
+### Added
+
+- Search heroes
+- Edit heroes
+- Snackbar
+- Uppercase Directive
+
 ## [0.0.1] - 2023-11-13
 
 ### Added